Created new component modules
